+++ src/api/java/blusunrize/immersiveengineering/api/Lib.java
@@ -76,6 +76,7 @@ public class Lib
 	public static final String GUIID_RedstoneConnector = "redstoneconnector";
 	public static final String GUIID_RedstoneProbe = "redstoneprobe";
 	public static final String GUIID_RedstoneStateCell = "redstonestatecell";
+	public static final String GUIID_RedstoneTimer = "redstonetimer";
 	//Items
 	public static final String GUIID_Revolver = "revolver";
 	public static final String GUIID_Toolbox = "toolbox";