+++ src/datagen/java/blusunrize/immersiveengineering/data/ItemModels.java
@@ -412,6 +412,9 @@ public class ItemModels extends TRSRItemModelProvider
 		obj(Connectors.REDSTONE_STATE_CELL, rl("block/connector/redstone_state_cell.obj"))
 				.transforms(rl("item/connector"))
 				.renderType(ModelProviderUtils.getName(translucent()));
+		obj(Connectors.REDSTONE_TIMER, rl("block/connector/redstone_timer.obj.ie"))
+				.transforms(rl("item/block"))
+				.renderType(ModelProviderUtils.getName(translucent()));
 		obj(Connectors.CONNECTOR_STRUCTURAL, rl("block/connector/connector_structural.obj.ie"))
 				.transforms(rl("item/connector"));
 		obj(Connectors.TRANSFORMER, rl("block/connector/transformer_mv_left.obj"))